chore: remove CoValueCore.fromID and CoValueCore.fromHeader and create the createUnloadedCoValue test utility
